+++ app/src/main/java/net/kullo/android/notifications/GcmConnector.java
@@ -8,6 +8,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.google.android.gms.common.ConnectionResult;
@@ -20,6 +21,14 @@ public class GcmConnector {
     private static final String TAG = "GcmConnector";
     private static final int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
 
+    public enum NotAvailableReason {
+        Disabled,
+        Missing,
+        UpdateRequired,
+        Updating,
+        Unknown,
+    }
+
     /* singleton setup */
     private static final GcmConnector SINGLETON = new GcmConnector();
     @NonNull public static GcmConnector get() {
@@ -27,7 +36,8 @@ public class GcmConnector {
     }
 
     /* members */
-    private Boolean mHasGooglePlay = null;
+    @Nullable private Boolean mGooglePlayOk = null;
+    @Nullable private NotAvailableReason mGooglePlayNotAvailableReason = null;
 
     // check gplay at app start, prompt user if not there
     @CheckResult
@@ -35,6 +45,24 @@ public class GcmConnector {
         GoogleApiAvailability playAPI = GoogleApiAvailability.getInstance();
         int result = playAPI.isGooglePlayServicesAvailable(activity);
         if (result != ConnectionResult.SUCCESS) {
+            switch (result) {
+                case ConnectionResult.SERVICE_DISABLED:
+                    mGooglePlayNotAvailableReason = NotAvailableReason.Disabled;
+                    break;
+                case ConnectionResult.SERVICE_MISSING:
+                    mGooglePlayNotAvailableReason = NotAvailableReason.Missing;
+                    break;
+                case ConnectionResult.SERVICE_VERSION_UPDATE_REQUIRED:
+                    mGooglePlayNotAvailableReason = NotAvailableReason.UpdateRequired;
+                    break;
+                case ConnectionResult.SERVICE_UPDATING:
+                    mGooglePlayNotAvailableReason = NotAvailableReason.Updating;
+                    break;
+                default:
+                    mGooglePlayNotAvailableReason = NotAvailableReason.Unknown;
+                    break;
+            }
+
             if (playAPI.isUserResolvableError(result)) {
                 return playAPI.getErrorDialog(activity, result, PLAY_SERVICES_RESOLUTION_REQUEST);
             } else {
@@ -42,9 +70,18 @@ public class GcmConnector {
                 return null;
             }
         } else {
-            mHasGooglePlay = true;
+            mGooglePlayOk = true;
             return null;
         }
+    }
+
+    public boolean googlePlayAvailable() {
+        RuntimeAssertion.require(mGooglePlayOk != null);
+        return mGooglePlayOk;
+    }
+
+    public NotAvailableReason googlePlayNotAvailableReason() {
+        return mGooglePlayNotAvailableReason;
     }
 
     // if conditions are right, launch service that will retrieve a new token
@@ -52,7 +89,8 @@ public class GcmConnector {
         RuntimeAssertion.require(SessionConnector.get().sessionAvailable());
         checkGooglePlay(context);
 
-        if (!mHasGooglePlay) {
+        RuntimeAssertion.require(mGooglePlayOk != null);
+        if (!mGooglePlayOk) {
             Log.i(TAG, "Device does not have Google Play Services. Skipping Push notifications.");
             return;
         }
@@ -73,10 +111,10 @@ public class GcmConnector {
     }
 
     private void checkGooglePlay(Context context) {
-        if (mHasGooglePlay == null) {
+        if (mGooglePlayOk == null) {
             GoogleApiAvailability playAPI = GoogleApiAvailability.getInstance();
             int result = playAPI.isGooglePlayServicesAvailable(context);
-            mHasGooglePlay = (result == ConnectionResult.SUCCESS);
+            mGooglePlayOk = (result == ConnectionResult.SUCCESS);
         }
     }
 }