+++ app/src/main/java/net/kullo/android/screens/SettingsActivity.java
@@ -39,6 +39,7 @@ import java.util.List;
 
 public class SettingsActivity extends AppCompatActivity {
     public static final String TAG = "SettingsActivity";
+    public static final int ACTIVITY_RESULT_PERMISSION_DENIED_READ_STORAGE = 8001;
 
     private static final int REQUEST_CODE_AVATAR_SOURCE_SELECTION = 1;
     private static final int REQUEST_CODE_CROP_AVATAR_IMAGE = 2;
@@ -155,7 +156,7 @@ public class SettingsActivity extends AppCompatActivity {
         mAvatarView.setOnClickListener(new View.OnClickListener() {
             //@Override
             public void onClick(View v) {
-                openSelectAvatarSourcesMenu();
+                openSelectAvatarSourceChooser();
             }
         });
     }
@@ -170,17 +171,16 @@ public class SettingsActivity extends AppCompatActivity {
         }
     }
 
-    private void openSelectAvatarSourcesMenu() {
-        // Determine Uri of camera image to save
-        final File cameraOutputDir = ((KulloApplication) getApplication()).captureCacheDir();
-        final File cameraOutputFile = new File(cameraOutputDir, "img_" + System.currentTimeMillis() + ".jpg");
-        mCameraOutputFileUri = FileProvider.getUriForFile(this, KulloApplication.ID, cameraOutputFile);
-        Log.d(TAG, "Camera output file: " + mCameraOutputFileUri);
+    private void openSelectAvatarSourceChooser() {
+        // Image picker (Gallery)
+        final Intent galleryIntent = new Intent(Intent.ACTION_PICK,
+                android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
+        galleryIntent.setType("image/");
 
-        // Option list
+        // Extras list
         final List<Intent> extraIntents = new ArrayList<>();
 
-        // Clear
+        // Extras: Clear
         final Intent clearIntent = new Intent();
         clearIntent.setClass(this, SettingsActivity.class);
         clearIntent.setAction(ACTION_CLEAR_AVATAR);
@@ -188,7 +188,14 @@ public class SettingsActivity extends AppCompatActivity {
             getString(R.string.settings_remove_avatar), R.drawable.kullo_settings_avatar);
         extraIntents.add(clearIntentOption);
 
-        // Camera
+        // Extras: Cameras
+        final File cameraOutputDir = ((KulloApplication) getApplication()).captureCacheDir();
+        final File cameraOutputFile = new File(cameraOutputDir, "img_" + System.currentTimeMillis() + ".jpg");
+        mCameraOutputFileUri = FileProvider.getUriForFile(this, KulloApplication.ID, cameraOutputFile);
+        Log.d(TAG, "Camera output file: " + mCameraOutputFileUri);
+
+        // Permission android.permission.CAMERA must not be declared when accessing camera via
+        // ACTION_IMAGE_CAPTURE intent. See http://stackoverflow.com/a/32856112/2013738
         final Intent captureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
         final PackageManager packageManager = this.getPackageManager();
         final List<ResolveInfo> cameras = packageManager.queryIntentActivities(captureIntent, 0);
@@ -198,18 +205,17 @@ public class SettingsActivity extends AppCompatActivity {
             intent.putExtra(MediaStore.EXTRA_OUTPUT, mCameraOutputFileUri);
             intent.setFlags(Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
             extraIntents.add(intent);
+
+            // Work around issue on older Android devices
+            // > Permission Denial: opening provider android.support.v4.content.FileProvider from ProcessRecord{[...] :com.android.hwcamera/[...]} that is not exported from uid [...]
+            // See http://stackoverflow.com/a/33652695/2013738
+            this.grantUriPermission(camera.activityInfo.packageName, mCameraOutputFileUri,
+                    Intent.FLAG_GRANT_WRITE_URI_PERMISSION | Intent.FLAG_GRANT_READ_URI_PERMISSION);
         }
 
-        // FileSystem
-        final Intent galleryIntent = new Intent(Intent.ACTION_PICK,
-                android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-        galleryIntent.setType("image/");
-
-        // Chooser of filesystem options
+        // Create and start avatar source chooser
         final Intent chooserIntent = Intent.createChooser(galleryIntent, getString(R.string.select_image_source));
         chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, extraIntents.toArray(new Parcelable[]{}));
-
-        // Launch the option menu
         startActivityForResult(chooserIntent, REQUEST_CODE_AVATAR_SOURCE_SELECTION);
     }
 
@@ -276,6 +282,9 @@ public class SettingsActivity extends AppCompatActivity {
                 updateAvatarViewFromSession();
             } else if (resultCode == Activity.RESULT_CANCELED) {
                 Toast.makeText(this, R.string.capture_image_canceled, Toast.LENGTH_SHORT).show();
+            } else if (resultCode == ACTIVITY_RESULT_PERMISSION_DENIED_READ_STORAGE) {
+                Toast.makeText(this, R.string.settings_permission_missing_storage,
+                        Toast.LENGTH_LONG).show();
             } else {
                 RuntimeAssertion.fail("Unhandled result code");
             }